[fix] : 테스트 코드 수정 (의존성 주입 방식으로 변경)

+++ src/main/java/atwoz/atwoz/common/auth/jwt/JwtProvider.java
@@ -20,8 +20,8 @@ public class JwtProvider {
     private final Key key;
 
     public JwtProvider(@Value("${jwt.secret}") String secret,
-                       @Value("${jwt.access-token.expiration}") int accessTokenExpiration,
-                       @Value("${jwt.refresh-token.expiration}") int refreshTokenExpiration) {
+                       @Value("${jwt.access-token.expiration}") long accessTokenExpiration,
+                       @Value("${jwt.refresh-token.expiration}") long refreshTokenExpiration) {
         this.key = Keys.hmacShaKeyFor(secret.getBytes());
         this.accessTokenExpiration = accessTokenExpiration;
         this.refreshTokenExpiration = refreshTokenExpiration;
